+++ protocols/mdns/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.46.0
+
+<!-- Update to libp2p-swarm v0.45.0 -->
 ## 0.45.2
 
 - Add `#[track_caller]` on all `spawn` wrappers.